work in progress
Former-commit-id: 2bf4527ffdc20f72aac635dd840a625a39329381
Former-commit-id: 70b7ce15b40fee933796e3af68b13cfa4cfd3073

+++ src/wid.h
@@ -132,8 +132,8 @@ typedef void(*on_destroy_t)(widp);
 typedef void(*on_tick_t)(widp);
 
 typedef widp (*grid_wid_replace_t)(widp,
-                                   int32_t x,
-                                   int32_t y,
+                                   double x,
+                                   double y,
                                    thing_templatep thing_template);
 
 /*
