+++ src/wid_editor_map.h
@@ -12,7 +12,7 @@ extern widp wid_editor_map_grid_container;
 extern boolean wid_editor_got_line_start;
 
 widp wid_editor_map_thing_replace_template(widp,
-                                           int32_t x, int32_t y,
+                                           double x, double y,
                                            thing_templatep);
 
 void wid_editor_map_thing_flood_fill_template(int32_t x, int32_t y,
